backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pydantic>=2.10
pydantic-settings>=2.7
python-dotenv>=1.0
cachetools>=5.5
openpyxl>=3.1
pytest>=8.0
httpx>=0.28
